:white_check_mark:(태훈) 유저 도메인 테스트 코드 추가

See merge request s09-bigdata-recom-sub2/S09P22A703!5

+++ backend/dksl/src/main/java/com/ssafy/dksl/model/dto/UserDto.java
@@ -35,7 +35,7 @@ public class UserDto {
     @NotNull
     private String puuid;
 
-    private List<Integer> teams;
+    private List<Long> teams;
 
 
 }
